# src/main/resources/openapi-doc.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/user": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithWishDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user-controller"
        ],
        "operationId": "updateUsername",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUsernameDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "upsertWishes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WishDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/exemptions": {
      "get": {
        "tags": [
          "exemptions-controller"
        ],
        "operationId": "getExemptionsForCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exemptions-controller"
        ],
        "operationId": "upsertExemptions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExemptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "exemptions-controller"
        ],
        "operationId": "deleteExemption",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExemptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/exemptions/all": {
      "get": {
        "tags": [
          "exemptions-controller"
        ],
        "operationId": "getAllExemptions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExemptionDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assignments/mine": {
      "get": {
        "tags": [
          "assignment-controller"
        ],
        "operationId": "getAssignmentForCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assignments/generate": {
      "get": {
        "tags": [
          "assignment-controller"
        ],
        "operationId": "generateAssignments",
        "responses": {
          "200": {
            "description": "OK",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChangeUsernameDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username"
        ]
      },
      "WishDto": {
        "type": "object",
        "properties": {
          "wish": {
            "type": "string"
          }
        },
        "required": [
          "wish"
        ]
      },
      "ExemptionRequest": {
        "type": "object",
        "properties": {
          "excludedUserId": {
            "type": "string"
          }
        },
        "required": [
          "excludedUserId"
        ]
      },
      "UserWithWishDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "wishes": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "wishes"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "username"
        ]
      },
      "ExemptionDto": {
        "type": "object",
        "properties": {
          "fromUser": {
            "$ref": "#/components/schemas/UserDto"
          },
          "toUser": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "fromUser",
          "toUser"
        ]
      }
    }
  }
}